+++ RedPandaIDE/translations/RedPandaIDE_pt_BR.ts
@@ -1452,12 +1452,12 @@
         <translation>Inserir a condição de parada:</translation>
     </message>
     <message>
-        <location line="+260"/>
+        <location line="+262"/>
         <source>Readonly</source>
         <translation>Apenas leitura</translation>
     </message>
     <message>
-        <location line="-5347"/>
+        <location line="-5349"/>
         <location line="+434"/>
         <source>Error Load File</source>
         <translation type="unfinished">Erro ao carregar arquivo</translation>
@@ -8374,7 +8374,7 @@
         <location filename="../autolinkmanager.cpp" line="+54"/>
         <location line="+16"/>
         <location line="+19"/>
-        <location filename="../settings.cpp" line="+4057"/>
+        <location filename="../settings.cpp" line="+4069"/>
         <location filename="../widgets/ojproblemsetmodel.cpp" line="-401"/>
         <location line="+61"/>
         <source>Can&apos;t open file &apos;%1&apos; for read.</source>
@@ -8689,7 +8689,7 @@
         <translation>Impossível carregar configurações para autolink</translation>
     </message>
     <message>
-        <location filename="../parser/cppparser.cpp" line="+1235"/>
+        <location filename="../parser/cppparser.cpp" line="+1258"/>
         <source>constructor</source>
         <translation>constructor</translation>
     </message>
@@ -11047,7 +11047,22 @@
         <translation>Limitar rolamento de página a apenas uma linha</translation>
     </message>
     <message>
-        <location line="+22"/>
+        <location line="+7"/>
+        <source>Mouse Scroll Direction</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location line="+6"/>
+        <source>Natural</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location line="+7"/>
+        <source>Invert</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location line="+38"/>
         <source>Mouse Wheel Scroll Speed</source>
         <translation>Velocidade de rolamento pela rodinha do mouse</translation>
     </message>